# [forge] Add metrics for insertion_to_block and block_creation_to_commit