## Description

Forge provides too much granularity with 4 columns capturing metrics for all parts the blockchain latency in 4 components. The goal of this PR is to simplify this into 2 columns capturing 

1. `InsertionToBlock` -> Time from getting into the node to getting into a block
2. `BlockCreationToCommit` -> Time from getting into the block to getting into storage

### OLD
![batch-pos](https://github.com/user-attachments/assets/64a4320c-cc7c-497b-a943-ffa7a94cd342)

### New

![Screenshot 2024-10-17 at 2 00 17 PM](https://github.com/user-attachments/assets/affd7c2c-b372-45b4-af37-80bc60960d9d)

## How Has This Been Tested?

Tested queries on grafana 👍 

- [Adhoc Forge Run before](https://github.com/aptos-labs/aptos-core/actions/runs/11264334213)
- [SUCCESS] [Adhoc Forge Run after (with `realistic_env_load_sweep`)](https://github.com/aptos-labs/aptos-core/actions/runs/11283913429)

## Evidence for latency limit changes (non-exhaustive)
- [FAILURE] [`Failed latency check, for \["P90 latency for 10 is 0.9s and exceeds limit of 0.9s"\]`](https://github.com/aptos-labs/aptos-core/actions/runs/11280108671)
- [FAILURE] [`Failed latency check, for \["P50 latency for 100 is 0.9s and exceeds limit of 0.9s"\]`](https://github.com/aptos-labs/aptos-core/actions/runs/11282998408)
- [FAILURE] [`Failed latency check, for \["P90 latency for 1000 is 1.3s and exceeds limit of 1.3s"\]`](https://github.com/aptos-labs/aptos-core/actions/runs/11282080946)