@hariria 2024-10-21
Moved to https://github.com/aptos-labs/aptos-core/pull/15036